Separate Windows specific code of NativeWindow to another file
